refactor: split e2ei enrolment from basic e2ei queries
